add bridge contract storage item and functionality
