+++ pallets/avn/src/lib.rs
@@ -18,7 +18,8 @@ extern crate alloc;
 #[cfg(not(feature = "std"))]
 use alloc::string::{String, ToString};
 
-use frame_support::{dispatch::DispatchResult, log::*, traits::OneSessionHandler};
+use frame_support::{dispatch::DispatchResult, ensure, log::*, traits::OneSessionHandler};
+use frame_system::{ensure_root, pallet_prelude::OriginFor};
 use sp_application_crypto::RuntimeAppPublic;
 use sp_avn_common::{
     event_types::{EthEventId, Validator},
@@ -26,6 +27,7 @@ use sp_avn_common::{
     recover_public_key_from_ecdsa_signature, DEFAULT_EXTERNAL_SERVICE_PORT_NUMBER,
     EXTERNAL_SERVICE_PORT_NUMBER_KEY,
 };
+use sp_core::{ecdsa, H160};
 use sp_runtime::{
     offchain::{http, storage::StorageValueRef, Duration},
     traits::Member,
@@ -36,11 +38,18 @@ use sp_std::prelude::*;
 use codec::{Decode, Encode};
 use core::convert::TryInto;
 pub use pallet::*;
-use sp_core::ecdsa;
 
 #[path = "tests/testing.rs"]
 pub mod testing;
 pub mod vote;
+
+#[cfg(test)]
+#[path = "tests/test_proxy_signed_add_avn_logs.rs"]
+mod test_proxy_signed_add_avn_logs;
+
+#[cfg(test)]
+#[path = "tests/test_set_bridge_contract.rs"]
+mod test_set_bridge_contract;
 
 // Definition of the crypto to use for signing
 pub mod sr25519 {
@@ -97,6 +106,7 @@ pub mod pallet {
         DeadlineReached,
         UnexpectedStatusCode,
         InvalidVotingSession,
+        InvalidContractAddress,
         DuplicateVote,
         InvalidVote,
         ErrorRecoveringPublicKeyFromSignature,
@@ -110,9 +120,41 @@ pub mod pallet {
     /// offchain worker.
     pub type Validators<T: Config> =
         StorageValue<_, Vec<Validator<T::AuthorityId, T::AccountId>>, ValueQuery>;
+
+    #[pallet::storage]
+    #[pallet::getter(fn get_bridge_contract_address)]
+    pub type AvnBridgeContractAddress<T: Config> = StorageValue<_, H160, ValueQuery>;
+
+    #[pallet::genesis_config]
+    pub struct GenesisConfig<T: Config> {
+        pub _phantom: sp_std::marker::PhantomData<T>,
+        pub bridge_contract_address: H160,
+    }
+
+    #[cfg(feature = "std")]
+    impl<T: Config> Default for GenesisConfig<T> {
+        fn default() -> Self {
+            Self { _phantom: Default::default(), bridge_contract_address: H160::zero() }
+        }
+    }
+
+    #[pallet::genesis_build]
+    impl<T: Config> GenesisBuild<T> for GenesisConfig<T> {
+        fn build(&self) {
+            AvnBridgeContractAddress::<T>::put(self.bridge_contract_address);
+        }
+    }
 }
 
 impl<T: Config> Pallet<T> {
+    pub fn set_bridge_contract(origin: OriginFor<T>, contract_address: H160) -> DispatchResult {
+        ensure_root(origin)?;
+        ensure!(&contract_address != &H160::zero(), Error::<T>::InvalidContractAddress);
+
+        <AvnBridgeContractAddress<T>>::put(contract_address);
+        Ok(())
+    }
+
     pub fn pre_run_setup(
         block_number: T::BlockNumber,
         caller_id: Vec<u8>,
@@ -541,7 +583,6 @@ impl<Balance> CollatorPayoutDustHandler<Balance> for () {
 }
 
 #[cfg(test)]
-#[path = "tests/mock.rs"]
 mod mock;
 
 #[cfg(test)]
